// components/runtime/src/ts/generated/internal/CustomEmoji.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface CustomEmoji {
  animated: boolean;
  available: boolean;
  id: string;
  managed: boolean;
  name: string;
  requireColons: boolean;
  roles: Array<string>;
  createdByUserId: string | null;
}